Move functions to get replica URLs
